Improved camera controls and math classes, made it c++11 compliant

+++ Engine/OpenGL/Primitive.h
@@ -32,7 +32,7 @@ private:
     };
     
     GLuint m_vboId, m_eboId, m_vaoId;
-    int m_time;
+    float m_time;
     
 };
